requirements.txt
torch==2.2.2
torchvision==0.17.2
ultralytics==8.1.30
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.1.post1
matplotlib==3.8.3
faiss-cpu==1.7.4
fastapi==0.110.0
uvicorn==0.27.1
plotly==5.19.0
tqdm==4.66.2
pyyaml==6.0.1
